[SKIP CI]fix terminal ns
